requirements.txt
# Python Dependencies for Robotics Scoring System

# Core Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0

# Serial Communication
pyserial>=3.5

# GUI
tk>=0.1.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (optional)
pylint>=2.17.0
black>=23.7.0

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0